step#4 - 메소드명 일치시키기, 도형 팩토리 클래스 추가, 추상메소드 리턴 타입 변경, 삼각형 넓이 구하기 구현(라인 활용, 포맷팅까지)

+++ src/main/java/domain/figure/Figure.java
@@ -9,7 +9,7 @@ public abstract class Figure {
         this.pointRepo = pointRepo;
     }
 
-    public abstract int calcArea();
+    public abstract double calcArea();
 
     public abstract String getFigureName();
 }
